% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{appReports}
\alias{appReports}
\title{Sample data for the shinymgr.sqlite table, "appReports"}
\format{
A data frame with 3 observations on the following 3 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
Reports are added to the shinymgr.sqlite database via the "Add
"Reports" menu in shinymgr's Developer section.
}
\section{Fields}{

\describe{
\item{\code{fkAppName}}{References an appName from the "apps" table.}

\item{\code{fkReportName}}{Reference a report from the "reports" table.}

\item{\code{notes}}{Notes about a report for a given app.}
}}

\examples{

# read in the demo data
data("demo_data")

# look at the structure
str(appReports)

}
\seealso{
Other data: 
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modFunctionReturns}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
