% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{appTabs}
\alias{appTabs}
\title{Sample data for the shinymgr.sqlite table, "appTabs"}
\format{
A data frame with 9 observations on the following 3 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

Records to the "appTabs" table are added to the shinymgr.sqlite 
database via the "App Builder" interface within shinymgr's Developer section.
}
\section{Fields}{

\describe{
\item{\code{fkTabName}}{References pkTabName in the "tabs" table.}

\item{\code{fkAppName}}{References pkAppName in the "apps" table.}

\item{\code{tabOrder}}{Specifies the order of tabs as presented to user.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(appTabs)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modFunctionReturns}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
