% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{modFunctionArguments}
\alias{modFunctionArguments}
\title{Sample data for the shinymgr.sqlite table, "modFunctionArguments"}
\format{
A data frame with 6 observations on the following 5 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

New records to "modFunctionArguments" table are added to the 
shinymgr.sqlite database via the  \code{\link{mod_register}} function.
}
\section{Fields}{

\describe{
\item{\code{pkModArgID}}{Auto-number primary key.}

\item{\code{fkModuleName}}{References pkModuleName in the "modules" table.}

\item{\code{functionArgName}}{Name of the argument.}

\item{\code{functionArgClass}}{Class of the argument (e.g., data.frame)}

\item{\code{description}}{Description of the argument.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(modFunctionArguments)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionReturns}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
