% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{modFunctionReturns}
\alias{modFunctionReturns}
\title{Sample data for the shinymgr.sqlite table, "modFunctionReturns"}
\format{
A data frame with 6 observations on the following 5 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

New records to "modFunctionReturns" table are added to the 
shinymgr.sqlite database via the  \code{\link{mod_register}} function.
}
\section{Fields}{

\describe{
\item{\code{pkModReturnID}}{References pkModuleName in the "modules" table.}

\item{\code{fkModuleName}}{References pkModuleName in the "modules" table.}

\item{\code{functionReturnName}}{Name of the return.}

\item{\code{functionReturnClass}}{Class of the return (e.g., data.frame)}

\item{\code{description}}{Description of the return.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(modFunctionReturns)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
