% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{modPackages}
\alias{modPackages}
\title{Sample data for the shinymgr.sqlite table, "modPackages"}
\format{
A data frame with 2 observations on the following 4 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

New records to "modPackages" table are added to the 
shinymgr.sqlite database via the  \code{\link{mod_register}} function.
}
\section{Fields}{

\describe{
\item{\code{fkModuleName}}{References pkModuleName in the "modules" table.}

\item{\code{packageName}}{Name of the R package}

\item{\code{version}}{Version of the package.}

\item{\code{notes}}{Notes about the package.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(modPackages)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modFunctionReturns}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
