% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{reports}
\alias{reports}
\title{Sample data for the shinymgr.sqlite table, "reports"}
\format{
A data frame with 3 observations on the following 3 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

Reports are added to the shinymgr.sqlite database via the "Add
"Reports" menu in shinymgr's Developer section.
}
\section{Fields}{

\describe{
\item{\code{pkReportName}}{Name of the report; primary key.}

\item{\code{displayName}}{Display name of the report.}

\item{\code{reportDescription}}{Description of the report.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(reports)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modFunctionReturns}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{tabModules}},
\code{\link{tabs}}
}
\concept{data}
