% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.R
\docType{data}
\name{tabs}
\alias{tabs}
\title{Sample data for the shinymgr.sqlite table, "tabs"}
\format{
A data frame with 10 observations on the following 4 variables.
}
\usage{
data(demo_data)
}
\description{
Sample data imported to the shinymgr SQLite database
 by the function \code{\link{shiny_db_populate}}.
}
\details{
-----------------------------------------------------------

New records to the "tabs" table are added to the shinymgr.sqlite 
database 
via the "App Builder" interface within shinymgr's Developer section.
}
\section{Fields}{

\describe{
\item{\code{pkTabName}}{The name of the tab; primary key.}

\item{\code{tabDisplayName}}{The name displayed on the tab.}

\item{\code{tabInstructions}}{Instructions for the tab.}

\item{\code{tabNotes}}{Notes on the tab.}
}}

\examples{

# read in the demo data
data(demo_data)

# look at the structure
str(tabs)

}
\seealso{
Other data: 
\code{\link{appReports}},
\code{\link{appStitching}},
\code{\link{appTabs}},
\code{\link{apps}},
\code{\link{modFunctionArguments}},
\code{\link{modFunctionReturns}},
\code{\link{modPackages}},
\code{\link{modules}},
\code{\link{reports}},
\code{\link{tabModules}}
}
\concept{data}
